Fixed a bug where the weight adjustment algo would inevitably break.
This bug was caused because of the if never updating closestDays, as closestDays is initialized with 0, it just stayed 0, hence resulting in a division by 0 error.

also fixed a error with the sql query in getRoom, where an unknown identifier (for that statement) has been used

adapted both queries to see an existing booking as invalid (to count towards weight), if the bookings end is smaller OR same as CURDATE(), instead of just before.
This has been done to take the change, that users aren't able to make a booking on the same day anymore, in account.